App.vue内のメイン部分をVue Routerによる表示に変更
